% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnote.R
\name{add_footnote_adv}
\alias{add_footnote_adv}
\title{Add advanced footnote}
\usage{
add_footnote_adv(kable_input, general = NULL, number = NULL,
  alphabet = NULL, symbol = NULL, footnote_order = c("general", "number",
  "alphabet", "symbol"), general_title = "Note: ", number_title = "",
  alphabet_title = "", symbol_title = "")
}
\description{
Add advanced footnote
}
