% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_rows.R
\name{auto_index}
\alias{auto_index}
\title{Automatically figuring out the group_row index}
\usage{
auto_index(x)
}
\arguments{
\item{x}{The index column. A vector. For example `c("a", "a", "b", "b", "b")``}
}
\description{
This helper function allows users to build the \code{group_row}
index more quickly and use \code{group_rows} in a way that is similar with
\code{collapse_rows}.
}
