% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_rows.R
\name{collapse_header_rows_html}
\alias{collapse_header_rows_html}
\title{Collapse repeated rows in table header to multirow table header cells}
\usage{
collapse_header_rows_html(kable_input, columns = NULL,
  valign = c("middle", "top", "bottom"))
}
\arguments{
\item{kable_input}{Output of \code{knitr::kable()} with \code{format} specified}

\item{columns}{A numeric value or vector indicating in which column(s) rows
need to be collapsed.}

\item{valign}{Select from "top", "middle"(default), "bottom". The reason why
"top" is not default is that the multirow package on CRAN win-builder is
not up to date.}
}
\value{
HTML table
}
\description{
Collapse repeated rows in table header to multirow table header cells
}
