% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_knitr.R
\name{escape_latex}
\alias{escape_latex}
\title{escape special LaTeX characters}
\usage{
escape_latex(x, newlines = FALSE, spaces = FALSE)
}
\description{
escape special LaTeX characters
}
\author{
Yihui Xie
}
