% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{html_dependency_bsTable}
\alias{html_dependency_bsTable}
\title{HTML dependency for Twitter bootstrap (table only)}
\usage{
html_dependency_bsTable()
}
\description{
HTML dependency for Twitter bootstrap (table only)
}
