% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{html_dependency_kePrint}
\alias{html_dependency_kePrint}
\title{HTML dependency for Javascript to enable bootstrap tooltip and popup message}
\usage{
html_dependency_kePrint()
}
\description{
HTML dependency for Javascript to enable bootstrap tooltip and popup message
}
