% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_add_header_above.R
\name{htmlTable_add_header_above}
\alias{htmlTable_add_header_above}
\title{Add an extra header row above the current header}
\usage{
htmlTable_add_header_above(kable_input, header = NULL)
}
\description{
Add an extra header row above the current header
}
