% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlTable_styling.R
\name{htmlTable_styling}
\alias{htmlTable_styling}
\title{HTML table attributes}
\usage{
htmlTable_styling(bootstrap_options = "bs-table", font_size = NULL,
  full_width = T)
}
\arguments{
\item{bootstrap_options}{Common HTML table formats}

\item{font_size}{}
}
\description{
This function provides a set of shortcuts to common HTML table
formats
}
