% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_themes.R
\name{kable_classic}
\alias{kable_classic}
\alias{kable_minimal}
\alias{kable_material}
\title{Alternative HTML themes}
\usage{
kable_classic(kable_input, striped = FALSE, hover = FALSE, ...)

kable_minimal(kable_input, striped = FALSE, hover = FALSE, ...)

kable_material(kable_input, striped = FALSE, hover = FALSE, ...)
}
\arguments{
\item{kable_input}{A HTML kable object.}

\item{striped}{T/F for adding striped rows.}

\item{hover}{T/F for adding hover effects.}

\item{...}{Everything else you need to specify in \code{kable_styling}.}
}
\description{
kableExtra uses the built-in bootstrap themes by default in
\code{kable_styling()}. Alternatively, you can use a customized table themes for
your table. This \code{lightable} table style sheet comes with three formats,
namely \code{lightable-minimal}, \code{lightable-classic} and \code{lightable-material} with
\code{hover} and \code{striped} options.
}
