% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{kable_to_parsed}
\alias{kable_to_parsed}
\alias{parsed_to_kable}
\title{Convert between kable text format and parsed format}
\usage{
kable_to_parsed(kable_input)

parsed_to_kable(parsed, kable_input)
}
\arguments{
\item{kable_input}{Output from \code{kbl()} or other functions producing
\code{"knitr_kable"} objects.}

\item{parsed}{Previous \link{LaTeX2} object to convert back.}
}
\value{
A \code{"LaTeX2"} object containing the parsed version of the
table.

A \code{"knitr_kable"} object corresponding to the \code{parsed} input.
}
\description{
Convert between kable text format and parsed format
}
