% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{landscape}
\alias{landscape}
\title{Print the table on an isolated landscape page in PDF}
\usage{
landscape(kable_input, margin = NULL)
}
\arguments{
\item{kable_input}{Output of `knitr::kable()` with `format` specified}

\item{margin}{Customizable page margin for special needs. Values can be
"1cm", "1in" or similar.}
}
\description{
For very wide tables in PDF documents, sometimes it might be a
good idea to have a single landscape page to house it. This function
will rotate the PDF page for the applied table in LaTeX environment. It won't
have any effects on HTML.
}
