% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linebreak.R
\name{linebreak}
\alias{linebreak}
\title{Make linebreak in LaTeX Table cells}
\usage{
linebreak(x, align = c("l", "c", "r"), double_escape = F, linebreaker = "\\n")
}
\arguments{
\item{x}{A character vector}

\item{align}{Choose from "l", "c" or "r".  Defaults to "l".}

\item{double_escape}{Whether special character should be double escaped.
Default is FALSE.}

\item{linebreaker}{Symbol for linebreaks to replace. Default is \verb{\\n}.}
}
\description{
This function generates LaTeX code of \code{makecell} so that users
can have linebreaks in their table
}
