% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_mirror.R
\name{magic_mirror}
\alias{magic_mirror}
\title{Magic mirror that returns kable's attributes}
\usage{
magic_mirror(input)
}
\arguments{
\item{input}{The output of kable}
}
\description{
Magic mirror that returns kable's attributes
}

