% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_mirror.R
\name{magic_mirror_html}
\alias{magic_mirror_html}
\title{Magic Mirror for html table --------}
\usage{
magic_mirror_html(input)
}
\arguments{
\item{input}{The output of kable}
}
\description{
Magic Mirror for html table --------
}
