% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_mirror.R
\name{magic_mirror_latex}
\alias{magic_mirror_latex}
\title{Magic mirror for latex tables --------------}
\usage{
magic_mirror_latex(kable_input)
}
\arguments{
\item{kable_input}{The output of kable}
}
\description{
Magic mirror for latex tables --------------
}
