% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helpers.R
\name{make_inline_plot}
\alias{make_inline_plot}
\title{Combine file (or svg text) and parameters into a \code{kableExtraInlinePlots} object}
\usage{
make_inline_plot(filename, file_ext, dev, width, height, res, del = TRUE)
}
\arguments{
\item{filename}{Passed through to the graphics device.}

\item{file_ext}{Character, something like "png".}

\item{dev}{Character (e.g., "svg", "pdf") or function (e.g.,}

\item{width, height}{Plot dimensions in pixels.}

\item{res}{The resolution of the plot; default is 300.}

\item{del}{If the file is svg-like, then the default action is to
read the file into an embedded SVG object; once done, the file is
no longer used. The default action is to delete this file early,
set this to 'FALSE' to keep the file.}
}
\value{
list object, with class \code{kableExtraInlinePlots}
}
\description{
Combine file (or svg text) and parameters into a \code{kableExtraInlinePlots} object
}
