% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_column.R
\name{remove_column}
\alias{remove_column}
\title{Remove columns}
\usage{
remove_column(kable_input, columns)
}
\arguments{
\item{kable_input}{Output of \code{\link[knitr:kable]{knitr::kable()}} with format specified}

\item{columns}{A numeric value or vector indicating in which column(s) rows
need to be removed}
}
\description{
Remove columns
}
\examples{
\dontrun{
remove_column(kable(mtcars), 1)
}
}
