% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_kable.R
\name{save_kable}
\alias{save_kable}
\title{Save kable to files}
\usage{
save_kable(x, file, bs_theme = "simplex", self_contained = TRUE)
}
\arguments{
\item{x}{A piece of HTML code for tables, usually generated by kable and
kableExtra}

\item{file}{save to files}

\item{bs_theme}{Which Bootstrap theme to use}

\item{self_contained}{Will the files be self-contained?}
}
\description{
Save kable to files
}
