% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_tools.R
\name{spec_angle}
\alias{spec_angle}
\title{Generate rotation angle for continuous values}
\usage{
spec_angle(x, begin, end)
}
\arguments{
\item{x}{continuous vectors of values}

\item{begin}{Smallest degree to rotate. Default is 0}

\item{end}{Largest degree to rotate. Default is 359.}
}
\description{
Generate rotation angle for continuous values
}
