% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_tools.R
\name{spec_font_size}
\alias{spec_font_size}
\title{Generate common font size for continuous values}
\usage{
spec_font_size(x, begin = 8, end = 16, na_font_size = 12, scale_from = NULL)
}
\arguments{
\item{x}{continuous vectors of values}

\item{begin}{Smallest font size to be used. Default is 10.}

\item{end}{Largest font size. Default is 20.}

\item{na_font_size}{font size for NA values}

\item{scale_from}{input range (vector of length two). If not given,
is calculated from the range of x}
}
\description{
Generate common font size for continuous values
}
