% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_tools.R
\name{spec_popover}
\alias{spec_popover}
\title{Setup bootstrap popover}
\usage{
spec_popover(
  content = NULL,
  title = NULL,
  trigger = "hover",
  position = "right"
)
}
\arguments{
\item{content}{content for pop-over message}

\item{title}{title for pop-over message.}

\item{trigger}{Controls how the pop-over message should be triggered.
Possible values include \code{hover} (default), \code{click}, \code{focus} and \code{manual}.}

\item{position}{How the tooltip should be positioned. Possible values are
\code{right}(default), \code{top}, \code{bottom}, \code{left} & \code{auto}.}
}
\description{
Setup bootstrap popover
}
