% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_tools.R
\name{spec_tooltip}
\alias{spec_tooltip}
\title{Setup bootstrap tooltip}
\usage{
spec_tooltip(title, position = "right")
}
\arguments{
\item{title}{text for hovering message}

\item{position}{How the tooltip should be positioned. Possible values are
\code{right}(default), \code{top}, \code{bottom}, \code{left} & \code{auto}.}
}
\description{
Setup bootstrap tooltip
}
