% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_header_above.R
\name{add_header_above}
\alias{add_header_above}
\title{Add an extra header row above the current header}
\usage{
add_header_above(kable_input, header = NULL)
}
\description{
Add an extra header row above the current header
}
