% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_header_left.R
\name{add_header_left}
\alias{add_header_left}
\title{Add a header column}
\usage{
add_header_left(kable_input, header = NULL, header_name = "")
}
\description{
Experimenting. Please don't use it in production
}
