% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_indent.R
\name{add_indent}
\alias{add_indent}
\title{Add indentations to row headers}
\usage{
add_indent(kable_input, positions)
}
\arguments{
\item{kable_input}{Output of `knitr::kable()` with `format` specified}

\item{positions}{A vector of numeric row numbers for the rows that need to
be indented.}
}
\description{
Add indentations to row headers
}
\examples{
x <- knitr::kable(head(mtcars), "html")
# Add indentations to the 2nd & 4th row
add_indent(x, c(2, 4))

}
