% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_spec.R
\name{cell_spec}
\alias{cell_spec}
\title{Specify Cell format}
\usage{
cell_spec(x, format, bold = F, italic = F, monospace = F, color = NULL,
  background = NULL, align = NULL, font_size = NULL, angle = NULL,
  hover_message = NULL, background_as_tile = TRUE)
}
\arguments{
\item{x}{Things to be formated. It could be a vector of numbers or strings.}

\item{format}{Either "html" or "latex". It can also be set through
\code{option(knitr.table.format)}, same as \code{knitr::kable()}.}

\item{bold}{A T/F value to control whether the text of the selected column
need to be bolded.}

\item{italic}{A T/F value to control whether the text of the selected column
need to be emphasized.}

\item{monospace}{A T/F value to control whether the text of the selected column
need to be monospaced (verbatim)}

\item{color}{A character string for column text color. Here please pay
attention to the differences in color codes between HTML and LaTeX.}

\item{background}{A character string for column background color. Here please
pay attention to the differences in color codes between HTML and LaTeX.}

\item{align}{A character string for cell alignment. For HTML, possible values could
be \code{l}, \code{c}, \code{r} plus \code{left}, \code{center}, \code{right}, \code{justify}, \code{initial} and \code{inherit}
while for LaTeX, you can only choose from \code{l}, \code{c} & \code{r}.}

\item{font_size}{Only if you want to specify font size locally in HTML.
This feature is not available in LaTeX}

\item{angle}{0-360, degree that the text will rotate. Can be a vector.}

\item{hover_message}{A vector of strings to be displayed as hover message.
Of course, this feature is nly available in HTML.}

\item{background_as_tile}{T/F value indicating if you want to have round
cornered tile as background.}
}
\description{
Specify Cell format before it gets into kable
}
