% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_rows.R
\name{group_rows}
\alias{group_rows}
\title{Put a few rows of a table into one category}
\usage{
group_rows(kable_input, group_name, start_row, end_row)
}
\description{
Put a few rows of a table into one category
}
