% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light_themes.R
\name{kable_classic}
\alias{kable_classic}
\alias{kable_classic_2}
\alias{kable_minimal}
\alias{kable_material}
\alias{kable_material_dark}
\alias{kable_paper}
\title{Alternative HTML themes}
\usage{
kable_classic(
  kable_input,
  lightable_options = "basic",
  html_font = "\\"Arial Narrow\\", \\"Source Sans Pro\\", sans-serif",
  ...
)

kable_classic_2(
  kable_input,
  lightable_options = "basic",
  html_font = "\\"Arial Narrow\\", \\"Source Sans Pro\\", sans-serif",
  ...
)

kable_minimal(
  kable_input,
  lightable_options = "basic",
  html_font = "\\"Trebuchet MS\\", verdana, sans-serif",
  ...
)

kable_material(
  kable_input,
  lightable_options = "basic",
  html_font = "\\"Source Sans Pro\\", helvetica, sans-serif",
  ...
)

kable_material_dark(
  kable_input,
  lightable_options = "basic",
  html_font = "\\"Source Sans Pro\\", helvetica, sans-serif",
  ...
)

kable_paper(
  kable_input,
  lightable_options = "basic",
  html_font = "\\"Arial Narrow\\", arial, helvetica, sans-serif",
  ...
)
}
\arguments{
\item{kable_input}{A HTML kable object.}

\item{lightable_options}{Options to customize lightable. Similar with
\code{bootstrap_options} in \code{kable_styling}. Choices include \code{basic}, \code{striped}
and \code{hover}.}

\item{html_font}{A string for HTML css font. For example,
\code{html_font = '"Arial Narrow", arial, helvetica, sans-serif'}.}

\item{...}{Everything else you need to specify in \code{kable_styling}.}
}
\description{
kableExtra uses the built-in bootstrap themes by default in
\code{kable_styling()}. Alternatively, you can use a customized table themes for
your table. This \code{lightable} table style sheet comes with three formats,
namely \code{lightable-minimal}, \code{lightable-classic}, \code{lightable-material} and
\code{lightable-material-dark} with \code{hover} and \code{striped} options.
}
