% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_styling.R
\name{kable_styling}
\alias{kable_styling}
\title{HTML table attributes}
\usage{
kable_styling(kable_input, bootstrap_options = "basic",
  latex_options = "basic", full_width = NULL, position = c("center",
  "left", "right", "float_left", "float_right"), font_size = NULL)
}
\arguments{
\item{bootstrap_options}{A character vector for bootstrap table options. For
detailed information, please check the package vignette or visit the
w3schools' \href{https://www.w3schools.com/bootstrap/bootstrap_tables.asp}{Bootstrap Page}
. Possible options include "basic", "striped", "bordered", "hover",
"condensed" and "responsive".}

\item{full_width}{A `TRUE` or `FALSE` variable controlling whether the HTML
table should have 100\% width.}

\item{position}{A character string determining whether and how the HTML table
should float on the page. Values could be "left", "center", "right"}

\item{font_size}{A numeric input for table font size}
}
\description{
This function provides a cleaner approach to modify the style
of HTML tables other than using the `table.attr` option in `knitr::kable()`.
Currenly, it assumes the HTML document has boot
}
