% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{kableExtra_latex_packages}
\alias{kableExtra_latex_packages}
\title{LaTeX Packages}
\usage{
kableExtra_latex_packages(xelatex = FALSE)
}
\arguments{
\item{xelatex}{Is \code{xelatex} going to be used to process
the file?}
}
\description{
This function shows all LaTeX packages that is supposed to be
loaded for this package in a R Markdown YAML format.
}
