% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{landscape}
\alias{landscape}
\title{Print the table on an isolated landscape page in PDF}
\usage{
landscape(kable_input, margin = NULL)
}
\arguments{
\item{kable_input}{Output of \code{knitr::kable()} with \code{format} specified}

\item{margin}{Customizable page margin for special needs. Values can be
"1cm", "1in" or similar.}
}
\description{
This function will put the table on an single landscape page.
It's useful for wide tables that can't be printed on a portrait page.
}
\examples{
\dontrun{
landscape(knitr::kable(head(mtcars), "latex"))
}

}
