% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_mirror.R
\name{magic_mirror}
\alias{magic_mirror}
\title{Magic mirror that returns kable's attributes}
\usage{
magic_mirror(kable_input)
}
\arguments{
\item{kable_input}{The output of kable}
}
\description{
Mirror mirror tell me, how does this kable look like?
}
\examples{
magic_mirror(knitr::kable(head(mtcars), "html"))
}
