% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scroll_box.R
\name{scroll_box}
\alias{scroll_box}
\title{Put a HTML table into a scrollable box}
\usage{
scroll_box(kable_input, height = NULL, width = NULL)
}
\arguments{
\item{kable_input}{A HTML kable object}

\item{height}{A character string indicating the height of the box, e.g. "50px"}

\item{width}{A character string indicating the width of the box, e.g. "100px"}
}
\description{
This function will put a HTML kable object in a fixed-height,
fixed-width or both box and make it scrollable.
}
