% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_tools.R
\name{spec_angle}
\alias{spec_angle}
\title{Generate rotation angle for continuous values}
\usage{
spec_angle(x, begin, end, scale_from = NULL)
}
\arguments{
\item{x}{continuous vectors of values}

\item{begin}{Smallest degree to rotate. Default is 0}

\item{end}{Largest degree to rotate. Default is 359.}

\item{scale_from}{input range (vector of length two). If not given,
is calculated from the range of x}
}
\description{
Generate rotation angle for continuous values
}
