% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{usepackage_latex}
\alias{usepackage_latex}
\title{Load a LaTeX package}
\usage{
usepackage_latex(name, options = NULL)
}
\arguments{
\item{name}{The LaTeX package name}

\item{options}{The LaTeX options for the package}
}
\description{
Load a LaTeX package using R code. Just like \verb{\\\\usepackage\{\}}
in LaTeX
}
\examples{
usepackage_latex("xcolor")
}
